CB-7734 - "navigator.notification.alert" or "navigator.notification.confirm" seem have a "many words" issue. (closes #39)
Changed plugin to use UIAlertController on iOS 8 and greater instead of
the UIAlertView.
It makes the message scrollable is the text is too long.

Fixes the crash on iOS 8.3 while fixing the too many works
problem on previous versions. The too many words problem is not present
on iOS 8.3

Fixes the problem on landscape mode.
The problem was still present on iPhone and iPod touch devices when the
dialog was shown on landscape mode, now it’s fixed

(cherry picked from commit ee3b02b3c78debe608263d847f7ac69bda1bfca0)